gcc compile fix (thanks to Heron for reporting the problem)
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@7184 3db68df2-c116-0410-a063-a993310a9797

+++ source/graphics/Terrain.h
@@ -80,7 +80,7 @@ public:
 	{
 		// GetSlope wants to access the next index,
 		// hence clamp to size-2
-		return clamp(index, 0, m_MapSize-2);
+		return clamp(index, ssize_t(0), m_MapSize-2);
 	}
 
 	float GetVertexGroundLevel(ssize_t i, ssize_t j) const;
